test(telegram): fold channel-push shape tests into test__bridge.py (fix PS-204)
CI's audit-project linter flagged tests/_telegram/test__channel_push.py
as an orphan: no matching src file under src/_telegram/. The bridge's
_process_update was the source it was testing, so move the five
channel-push payload-shape tests into test__bridge.py (one per
assertion, AAA structure preserved) and delete the orphan file.

Behaviour-equivalent: same five tests, same fixtures, just a different
file.
